requirements.txt
flask
passlib
flask_jwt_extended 
flask_sqlalchemy
python-dotenv
marshmallow
flask_smorest